requirements.txt
pika==1.2.1
PyYAML==6.0
ecs_logging==2.0.0